Customer Rating Analysis
This dashboard consists of 3 graphs which are based on the data related to customer reviews/ feedback on Amazon.com and Walmart.com. 

The first graph analyses the count and percentage of stars given by customers

The second graph ( Positive feedback) gives a count of the number of stars given by customers whose feedback was categorized as "positive" with the x-axis showing the main words used in the comment.

The third graph ( Critical feedback) gives a count of the number of stars given by customers whose feedback was categorized as "critical" with the x-axis showing the main words used in the comment.